Expanded on the design decision to let the main thread close all its queues with a 2-step closing process.
