+++ src/wenum/core.py
@@ -3,11 +3,17 @@ from __future__ import annotations
 import logging
 from typing import TYPE_CHECKING
 
+from typing import Optional
+from threading import RLock
+
 if TYPE_CHECKING:
     from wenum.runtime_session import FuzzSession
 from .fuzzobjects import FuzzType, FuzzItem
 
-from .myqueues import MyPriorityQueue, QueueManager
+import collections
+from itertools import zip_longest
+
+from .myqueues import MyPriorityQueue, FuzzQueue, MonitorFuzzQueue
 from .fuzzqueues import (
     SeedQueue,
     FilePrinterQ,
@@ -113,6 +119,7 @@ class Fuzzer:
         fuzz_item: FuzzItem = self.last_queue.get()
         self.last_queue.task_done()
 
+        # STOP item may come from the monitorqueue. Raise an according exception in that case
         if fuzz_item.item_type == FuzzType.STOP:
             self.logger.debug("core.py: Stopping the queues")
             raise StopIteration
@@ -122,14 +129,19 @@ class Fuzzer:
 
         return fuzz_item
 
-    def stats(self):
+    def stats(self) -> dict:
         return dict(
             list(self.qmanager.get_stats().items())
             + list(self.qmanager["transport_queue"].http_pool.job_stats().items())
             + list(self.session.compiled_stats.get_runtime_stats().items())
         )
 
-    def cancel_job(self):
+    def cancel_job(self) -> None:
+        """
+        Method called when an exception of some sort is catched by the main loop.
+        If no requests are left or if the user sends an interrupt, it also ends up throwing an exception that
+        leads to the execution of this method.
+        """
         self.qmanager.stop_queues()
 
     def pause_job(self):
@@ -137,3 +149,161 @@ class Fuzzer:
 
     def resume_job(self):
         self.qmanager["transport_queue"].pause.set()
+
+
+class QueueManager:
+    """
+    Class responsible for keeping track of all the active queues and managing them
+    """
+    def __init__(self, session):
+        self._queues: collections.OrderedDict[str, FuzzQueue] = collections.OrderedDict()
+        # Queue at the end of the chain to e.g. check if all requests are done
+        self.monitor_queue: Optional[MonitorFuzzQueue] = None
+        # Queue receiving information from monitor_queue; last_queue items will be pulled by the main thread
+        self.last_queue: Optional[MyPriorityQueue] = None
+        self._mutex = RLock()
+        self.logger = logging.getLogger("debug_log")
+
+        self.session = session
+
+    def add(self, name, queue):
+        """
+        Add another Queue to the manager
+        """
+        self._queues[name] = queue
+
+    def move_to_end(self, key, last=True):
+        """
+        Execute move_to_end function of OrderedDict
+        """
+        try:
+            self._queues.move_to_end(key, last)
+        except KeyError:
+            raise
+
+    def get_stats(self):
+        stat_list = []
+
+        for queue in list(self._queues.values()):
+            stat_list = stat_list + list(queue.get_stats().items())
+
+        return dict(stat_list)
+
+    def bind(self, last_queue: MyPriorityQueue):
+        """Set all the correct output queues."""
+        with self._mutex:
+            self.last_queue = last_queue
+
+            queue_list: list[FuzzQueue] = list(self._queues.values())
+
+            self.monitor_queue: MonitorFuzzQueue = MonitorFuzzQueue(self.session, last_queue)
+
+            # Set the next queue for each queue
+            for index, (first, second) in enumerate(zip_longest(queue_list[0:-1], queue_list[1:])):
+                first.next_queue(second)
+                first.set_syncq(self.monitor_queue)
+                # Check for the remaining next queues if they are processing discarded fuzzresults
+                for next_one in queue_list[index + 1:]:
+                    if next_one.process_discarded:
+                        first.queue_discard = next_one
+                        break
+                # If none of the remaining queues intend to process discarded results,
+                # set the sync as the discard queue
+                if not first.queue_discard:
+                    first.queue_discard = self.monitor_queue
+
+            # The last queue receives the monitor queue as it's next queue
+            queue_list[-1].next_queue(self.monitor_queue)
+            queue_list[-1].set_syncq(self.monitor_queue)
+            queue_list[-1].queue_discard = self.monitor_queue
+
+    def __getitem__(self, key):
+        return self._queues[key]
+
+    def join(self, remove=False):
+        with self._mutex:
+            for k, queue in list(self._queues.items()):
+                self.logger.debug(f"Joining queue {queue.get_name()}")
+                queue.join()
+                self.logger.debug(f"Joined queue {queue.get_name()}")
+                if remove:
+                    self.logger.debug(f"Removing queue {queue.get_name()}")
+                    del self._queues[k]
+                    self.logger.debug(f"Removed queue {queue.get_name()}")
+
+    def start(self):
+        """
+        Starting method called by the core
+        """
+        with self._mutex:
+            if self._queues:
+                self.monitor_queue.qstart()
+                for queue in list(self._queues.values()):
+                    queue.qstart()
+
+                list(self._queues.values())[0].put_important(FuzzItem(FuzzType.STARTSEED))
+
+    def cleanup(self):
+        """
+        Called to end the runtime#TODO about to be phased out by stop_queues()
+        """
+        with self._mutex:
+            if self._queues:
+                self.join(remove=True)
+                self.last_queue.put_important(None, block=False)
+
+                self._queues = collections.OrderedDict()
+                self.last_queue = None
+        self.logger.debug("QueueManager: Cleaned up.")
+
+    def stop_queues(self):
+        """
+        Called to stop all active queues.
+
+        This is a 2-step process of first stopping their main loop and then
+        actually closing. This prevents racing conditions. Queues may send each other items, which will
+        prevent joining if a queue isn't empty when joining is attempted. The aforementioned 2-step process ensures
+        that such a scenario will not happen.
+        """
+        with self._mutex:
+            if self._queues:
+                self.logger.debug("QueueManager: Closing all queues")
+                # Send signal to all queues to stop their main loop
+                for active_queue in list(self._queues.values()):
+                    #TODO All queues need to handle this now. They should do all the cancelling themselves instead
+                    # of the main thread doing it for them if possible
+                    # To be able to join, the queues need to be empty.
+                    active_queue.cancel()
+                    active_queue.put_important(FuzzItem(FuzzType.STOP))
+                    self.logger.debug(f"QueueManager: Sent stop signal to {active_queue.get_name()}")
+                self.monitor_queue.put_important(FuzzItem(FuzzType.STOP))
+                self.logger.debug(f"QueueManager: Sent stop signal to {self.monitor_queue.name}")
+
+                # Wait until all queues confirmed that they have stopped their main loop
+                for active_queue in list(self._queues.values()):
+                    active_queue.stopped.wait()
+                self.monitor_queue.stopped.wait()
+                self.logger.debug(f"QueueManager: All queues have stopped")
+
+                # Send signal to all queues to ensure they are ready to join their thread
+                for active_queue in list(self._queues.values()):
+                    active_queue.close.set()
+
+                # Close all queue threads
+                for active_queue in list(self._queues.values()):
+                    self.logger.debug(f"QueueManager joining queue {active_queue.get_name()}")
+                    active_queue.join()
+                self.logger.debug(f"QueueManager joining queue {self.monitor_queue.name}")
+                self.monitor_queue.join()
+
+                self.logger.debug(f"QueueManager: joining lastqueue")
+
+                # Finally stop the last queue
+                while self.last_queue.qsize() > 0:
+                    self.last_queue.get()
+                    self.last_queue.task_done()
+                self.last_queue.join()
+
+                self.logger.debug("QueueManager: All queues have joined")
+
+                #self.cleanup()
